Node Manager, Node Tester and JSDoc

Approved-by: Thierry Parent <tparent@absolunet.com>

+++ test/unit/cache/services/CacheManager/drivers/FileDriver.gwt.js
@@ -1,13 +1,13 @@
 //--------------------------------------------------------
 //-- Tests - Unit - Cache - Services - Cache Manager - Drivers - File Driver - GWT
 //--------------------------------------------------------
-'use strict';
 
-const { given, when, then, build } = require('./common.gwt');
+import gwt from './common.gwt';
+const { given, when, then, build } = gwt;
 
-const fs         = require('fs');
-const path       = require('path');
-const FileDriver = require('../../../../../../lib/cache/services/CacheManager/drivers/FileDriver');
+import * as fs    from 'fs';
+import * as path  from 'path';
+import FileDriver from '../../../../../../dist/node/cache/services/CacheManager/drivers/FileDriver';
 
 const fileDriverCacheFolderPath = path.join(__dirname, '..', '..', '..', 'stubs', 'data');
 const fileDriverCacheFilePath   = path.join(fileDriverCacheFolderPath, 'test.json');
@@ -24,7 +24,9 @@ given.fileDriver = (parameters) => {
 };
 
 given.existingCacheDirectory = () => {
-	fs.mkdirSync(fileDriverCacheFolderPath);
+	if (!fs.existsSync(fileDriverCacheFolderPath)) {
+		fs.mkdirSync(fileDriverCacheFolderPath);
+	}
 };
 
 
@@ -38,8 +40,10 @@ then.removeCacheFile = () => {
 };
 
 then.removeCacheDirectory = () => {
-	fs.rmdirSync(fileDriverCacheFolderPath);
+	if (fs.existsSync(fileDriverCacheFolderPath)) {
+		fs.rmdirSync(fileDriverCacheFolderPath);
+	}
 };
 
 
-module.exports = build({ given, when, then });
+export default build({ given, when, then });
